Move ParsingMode into style_traits.
